requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-classic
langchain-core
langchain-text-splitters
langchain-mongodb
pymongo
sentence-transformers
python-dotenv
